A working version with no 'state machine' for now. Added new functions to usart for uint32 and half uint32. Changed uint and i2c_safe init such that speed/baude and f_cpu are parameters. Changed the Makefile to avoid linker errors with -lm

+++ i2c_safe.c
@@ -21,7 +21,7 @@
 /*************************************************************************
  I nitialization* of the I2C bus interface. Need to be called only once
  *************************************************************************/
-void i2c_init(uint16_t I2Cfreq, uint16_t fcpu)
+void i2c_init(uint32_t I2Cfreq, uint16_t fcpu)
 {
     /* initialize TWI clock: 100 kHz clock, TWPS = 0 => prescaler = 1 */
